add controls to jsf view for controlling log generation; app can now write to multiple logfiles simultaneously and each logfile definition will have its own thread for log generation

+++ src/main/java/se/kits/stuff/model/LogFileDefinition.java
@@ -18,6 +18,6 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class LogFileDefinition {
     private String fileName;
     private String logPattern;
-    private int timeSkewSeconds;
+    private double timeSkewSeconds;
     private double frequencyPerMinute;
 }
